* dda/pr/8547: (44 commits)
  True last sign
  UI signs
  Last signs
  Other signs
  Vehicle signs
  Map signs
  iuse signs
  pickup signs
  Ranged signs
  Item signs
  NPCtalk signs and restyles
  output signs
  NPCmove signs
  Autopickup signs
  Options signs
  Inventory signs
  iexamine signs
  Monster signs
  Construction signs
  Monmove signs
  ...

+++ src/action.cpp
@@ -711,14 +711,14 @@ action_id handle_action_menu()
         int width = 0;
         for (std::vector<uimenu_entry>::iterator entry = entries.begin();
              entry != entries.end(); ++entry) {
-            if (width < entry->txt.length()) {
+            if (width < (int)entry->txt.length()) {
                 width = entry->txt.length();
             }
         }
         //border=2, selectors=3, after=3 for balance.
         width += 2 + 3 + 3;
         int ix = (TERMX > width) ? (TERMX - width) / 2 - 1 : 0;
-        int iy = (TERMY > entries.size() + 2) ? (TERMY - entries.size() - 2) / 2 - 1 : 0;
+        int iy = (TERMY > (int)entries.size() + 2) ? (TERMY - (int)entries.size() - 2) / 2 - 1 : 0;
         int selection = (int) uimenu(true, std::max(ix, 0), std::min(width, TERMX - 2),
                                      std::max(iy, 0), title, entries);
