+++ src/defense.cpp
@@ -34,7 +34,7 @@ std::vector<int> original_recipe_values;
 int caravan_price(player &u, int price);
 
 void draw_caravan_borders(WINDOW *w, int current_window);
-void draw_caravan_categories(WINDOW *w, int category_selected, int total_price,
+void draw_caravan_categories(WINDOW *w, int category_selected, unsigned total_price,
                              unsigned long cash);
 void draw_caravan_items(WINDOW *w, std::vector<itype_id> *items,
                         std::vector<int> *counts, int offset,
@@ -975,7 +975,7 @@ void defense_game::caravan()
         }
     }
 
-    int total_price = 0;
+    unsigned total_price = 0;
 
     WINDOW *w = newwin(FULL_SCREEN_HEIGHT, FULL_SCREEN_WIDTH, 0, 0);
 
@@ -1027,7 +1027,7 @@ Press %s to buy everything in your cart, %s to buy nothing."),
                                    item_selected);
                 draw_caravan_borders(w, current_window);
             } else if (!items[category_selected].empty()) { // Items
-                if (item_selected < items[category_selected].size() - 1) {
+                if (item_selected < (int)items[category_selected].size() - 1) {
                     item_selected++;
                 } else {
                     item_selected = 0;
@@ -1335,7 +1335,7 @@ void draw_caravan_borders(WINDOW *w, int current_window)
     wrefresh(w);
 }
 
-void draw_caravan_categories(WINDOW *w, int category_selected, int total_price,
+void draw_caravan_categories(WINDOW *w, int category_selected, unsigned total_price,
                              unsigned long cash)
 {
     // Clear the window
@@ -1364,7 +1364,7 @@ void draw_caravan_items(WINDOW *w, std::vector<itype_id> *items,
         mvwprintz(w, i, 1, c_black, "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
     }
     // THEN print it--if item_selected is valid
-    if (item_selected < items->size()) {
+    if (item_selected < (int)items->size()) {
         item tmp( (*items)[item_selected] , 0); // Dummy item to get info
         fold_and_print(w, 12, 1, 38, c_white, tmp.info());
     }
@@ -1373,12 +1373,12 @@ void draw_caravan_items(WINDOW *w, std::vector<itype_id> *items,
         mvwprintz(w, i, 40, c_black, "xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx");
     }
     // Finally, print the item list on the right
-    for (int i = offset; i <= offset + FULL_SCREEN_HEIGHT - 2 && i < items->size(); i++) {
+    for (int i = offset; i <= offset + FULL_SCREEN_HEIGHT - 2 && i < (int)items->size(); i++) {
         mvwprintz(w, i - offset + 1, 40, (item_selected == i ? h_white : c_white),
                   itypes[ (*items)[i] ]->nname((*counts)[i]).c_str());
         wprintz(w, c_white, " x %2d", (*counts)[i]);
         if ((*counts)[i] > 0) {
-            int price = caravan_price(g->u, itypes[(*items)[i]]->price * (*counts)[i]);
+            unsigned price = caravan_price(g->u, itypes[(*items)[i]]->price * (*counts)[i]);
             wprintz(w, (price > g->u.cash ? c_red : c_green), "($%6d)", price);
         }
     }
